+++ plottool_ibeis/__main__.py
@@ -1,6 +1,4 @@
 #!/usr/bin/env python
-# -*- coding: utf-8 -*-
-from __future__ import absolute_import, division, print_function
 
 
 def plottool_ibeis_main():