+++ plottool_ibeis/tests/test_helpers.py
@@ -1,4 +1,3 @@
-from __future__ import absolute_import, division, print_function
 
 
 def dummy_bbox(img, shiftxy=(0.0, 0.0), scale=.25):