+++ plottool_ibeis/tests/test_viz_images.py
@@ -1,5 +1,4 @@
 #!/usr/bin/env python
-from __future__ import absolute_import, division, print_function
 from plottool_ibeis import viz_image2
 from plottool_ibeis import draw_func2 as df2
 from plottool_ibeis import plot_helpers as ph